+++ automated-tests/src/dali/utc-Dali-GeoHoverProcessing.cpp
@@ -1016,7 +1016,7 @@ int UtcDaliGeoHoverMultipleRenderTasks(void)
   scene.Add(actor);
 
   // Create render task
-  Viewport   viewport(sceneSize.width * 0.5f, sceneSize.height * 0.5f, sceneSize.width * 0.5f, sceneSize.height * 0.5f);
+  Viewport   viewport(static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f), static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f));
   RenderTask renderTask(application.GetScene().GetRenderTaskList().CreateTask());
   renderTask.SetViewport(viewport);
   renderTask.SetInputEnabled(true);
@@ -1067,7 +1067,7 @@ int UtcDaliGeoHoverMultipleRenderTasksWithChildLayer(void)
   actor.Add(layer);
 
   // Create render task
-  Viewport   viewport(sceneSize.width * 0.5f, sceneSize.height * 0.5f, sceneSize.width * 0.5f, sceneSize.height * 0.5f);
+  Viewport   viewport(static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f), static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f));
   RenderTask renderTask(application.GetScene().GetRenderTaskList().CreateTask());
   renderTask.SetViewport(viewport);
   renderTask.SetInputEnabled(true);
@@ -1110,7 +1110,7 @@ int UtcDaliGeoHoverOffscreenRenderTasks(void)
   Vector2                  sceneSize(scene.GetSize());
 
   // FrameBufferImage for offscreen RenderTask
-  FrameBuffer frameBuffer = FrameBuffer::New(sceneSize.width, sceneSize.height);
+  FrameBuffer frameBuffer = FrameBuffer::New(static_cast<uint32_t>(sceneSize.width), static_cast<uint32_t>(sceneSize.height));
 
   // Create a renderable actor to display the FrameBufferImage
   Actor renderableActor = CreateRenderableActor(frameBuffer.GetColorTexture());
